Refactor: 홈 화면 추가 모듈화
